requirements.txt
openai
python-dotenv
tqdm
ipywidgets
tiktoken
pypandoc
openai-agents==0.2.9
streamlit
markdown
pandarallel
openpyxl
weave
attrs
weave[all]
tavily-python
exa-py
plotly
pandas
numpy
seaborn
matplotlib
streamlit
galileo
xlrd==1.2.0